+++ CHANGELOG.md
@@ -24,3 +24,4 @@
     - Visitor Service
     - Remote Commands
     - Hosted Data Layer
+    - Crash Reporter